backend/requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-chroma
langchain-text-splitters
langchain-huggingface
langchain-groq
chromadb
pypdf
sentence-transformers
torch
groq
fastapi
uvicorn
motor
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
openai
langchain-openai
